+++ src/modules/layerInformation/store/actionsLayerInformation.js
@@ -113,6 +113,8 @@ export default {
             commit("setDatePublication", "");
             commit("setAbstractText", i18next.t("common:modules.layerInformation.noMetadataLoaded"));
             commit("setNoMetadataLoaded", i18next.t("common:modules.layerInformation.noMetadataLoaded"));
+            commit("setPointOfContact", "");
+            commit("setPublisher", "");
         }
         else {
             commit("setTitle", metadata?.getTitle());
@@ -120,6 +122,8 @@ export default {
             commit("setPeriodicityKey", metadata?.getFrequenzy());
             commit("setDownloadLinks", metadata?.getDownloadLinks());
             commit("setDatePublication", metadata?.getPublicationDate() || metadata?.getCreationDate());
+            commit("setPointOfContact", metadata?.getContact());
+            commit("setPublisher", metadata?.getPublisher());
         }
 
         if (state.downloadLinks) {